PB-1439: Remove duplicate fixtures
