+++ app/tests/access/test_access_models.py
@@ -3,26 +3,11 @@ from unittest.mock import patch
 
 from access.models import CognitoInconsistencyError
 from access.models import User
-from provider.models import Provider
-from pytest import fixture
 from pytest import raises
 
 from django.core.exceptions import ValidationError
 from django.forms import ModelForm
 from django.utils import timezone
-
-
-@fixture(name='provider')
-def fixture_provider(db):
-    yield Provider.objects.create(
-        slug="ch.bafu",
-        acronym_de="BAFU",
-        acronym_fr="OFEV",
-        acronym_en="FOEN",
-        name_de="Bundesamt für Umwelt",
-        name_fr="Office fédéral de l'environnement",
-        name_en="Federal Office for the Environment"
-    )
 
 
 @patch('access.models.Client')
